refactor: consolidate peaking EQ node state
